+++ packages/core/src/atomic-rule.ts
@@ -19,9 +19,11 @@ export type ProcessOptions = {
 
 export class AtomicRule {
   root: Root
-  layer = 'utilities'
+  layer: string
+
   constructor(private context: StylesheetContext) {
     this.root = postcss.root()
+    this.layer = context.layers.utilities
   }
 
   hashFn = (conditions: string[], className: string) => {