+++ packages/generator/src/artifacts/css/flat-css.ts
@@ -10,7 +10,7 @@ export const generateFlattenedCss = (ctx: Context) => (options: { files: string[
   const { theme: { keyframes } = {}, preflight, minify, staticCss } = ctx.config
 
   const unresolved = [
-    '@layer reset, base, tokens, recipes, utilities;',
+    ctx.layerString,
     preflight && "@import './reset.css';",
     "@import './global.css';",
     staticCss && "@import './static.css';",