+++ packages/core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pandacss/core
 
+## 0.11.1
+
+### Patch Changes
+
+- 23b516f4: Make layers customizable
+- Updated dependencies [c07e1beb]
+- Updated dependencies [23b516f4]
+  - @pandacss/shared@0.11.1
+  - @pandacss/types@0.11.1
+  - @pandacss/token-dictionary@0.11.1
+  - @pandacss/error@0.11.1
+  - @pandacss/logger@0.11.1
+
 ## 0.11.0
 
 ### Patch Changes