+++ packages/node/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @pandacss/node
 
+## 0.11.1
+
+### Patch Changes
+
+- 23b516f4: Make layers customizable
+- Updated dependencies [c07e1beb]
+- Updated dependencies [dfb3f85f]
+- Updated dependencies [23b516f4]
+  - @pandacss/generator@0.11.1
+  - @pandacss/shared@0.11.1
+  - @pandacss/is-valid-prop@0.11.1
+  - @pandacss/types@0.11.1
+  - @pandacss/core@0.11.1
+  - @pandacss/parser@0.11.1
+  - @pandacss/token-dictionary@0.11.1
+  - @pandacss/config@0.11.1
+  - @pandacss/error@0.11.1
+  - @pandacss/extractor@0.11.1
+  - @pandacss/logger@0.11.1
+
 ## 0.11.0
 
 ### Patch Changes