+++ packages/postcss/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pandacss/postcss
 
+## 0.11.1
+
+### Patch Changes
+
+- Updated dependencies [23b516f4]
+  - @pandacss/node@0.11.1
+
 ## 0.11.0
 
 ### Patch Changes