+++ packages/token-dictionary/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pandacss/token-dictionary
 
+## 0.11.1
+
+### Patch Changes
+
+- Updated dependencies [c07e1beb]
+- Updated dependencies [23b516f4]
+  - @pandacss/shared@0.11.1
+  - @pandacss/types@0.11.1
+
 ## 0.11.0
 
 ### Patch Changes